requirements.txt
pytest
pytest-mock
pytest-cov
pytest-randomly
pyhamcrest
requests-mock
coverage
